+++ src/main/java/it/univaq/esc/repository/AppuntamentoRepository.java
@@ -11,5 +11,5 @@ public interface AppuntamentoRepository extends JpaRepository<Appuntamento, Inte
 
     //public List<Appuntamento> findByPrenotazioneSpecsAppuntamento_ImpiantoPrenotato(Impianto impiantoPrenotato);
 
-    public Appuntamento findByPrenotazioneSpecsAppuntamento_Id(Long idPrenotazioneSPecs);
+    public Appuntamento findByPrenotazioneSpecsAppuntamento_Id(Integer idPrenotazioneSPecs);
 }
